airflow/requirements.txt
python-dateutil==2.8.1
kafka-python==2.0.1
redis==3.5.3
requests==2.23.0
fakeredis==1.4.1
pytest==5.4.3
python-coveralls==2.9.3
pytest==5.4.3
pytest-cov==2.10.1
coveralls==2.1.2
tweepy
retry
twint
vaderSentiment